+++ spec/rubocop/cop/rspec/file_path_spec.rb
@@ -57,6 +57,13 @@ RSpec.describe RuboCop::Cop::RSpec::FilePath do
     RUBY
   end
 
+  it 'registers an offense when second argument contains spaces' do
+    expect_offense(<<-RUBY, 'wrong_class_spec.rb')
+      describe MyClass, "via `local_failures`" do; end
+      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Spec path should end with `my_class*via_local_failures*_spec.rb`.
+    RUBY
+  end
+
   it 'registers an offense for a file missing _spec' do
     expect_offense(<<-RUBY, 'spec/models/user.rb')
       describe User do; end